Repo Cleanup
- Updated .env.example to include configurations for both local and Sepolia environments, providing clearer examples for users.
- Removed the obsolete .env.sepolia file to streamline environment management.
- Enhanced the CacheManagerAutomation contract by removing the deprecated AutomationCompatibleInterface and updating internal logic for better performance.
- Updated hardhat.config.ts and foundry.toml to reflect new build paths and configurations.
- Cleaned up test files and removed unused scripts to improve project organization.
