+++ .env.example
@@ -1,12 +1,29 @@
-ENVIRONMENT=local
+# LOCAL EXAMPLE - BASED ON NITRO TESTNODE
 CACHE_MANAGER_ADDRESS=0x0f1f89aaf1c6fdb7ff9d361e4388f5f3997f12a8
 ARB_WASM_CACHE_ADDRESS=0x0000000000000000000000000000000000000072
+CMA_ADDRESS="[DEPLOY AND DEFINE]"
 ARBPRE_ADD=0x3f1Eae7D46d88F08fc2F8ed27FCb2AB183EB2d0E
 ARBPRE_PK=0xb6b15c8cb491557369f3c7d2c287b053eb229daa9c22138887752191c9520659
 ARBLOC_OWNER_ADD=0x5E1497dD1f08C87b2d8FE23e9AAB6c1De833D927
 ARBLOC_OWNER_PK=0xdc04c5399f82306ec4b4d654a342f40e2e0620fe39950d967e1e574b32d4dd36
+USER_ADD="[ANY ADDRESS WITH FUNDS]"
+USER_PK="[ANY ADDRESS WITH FUNDS]"
+USER_ADD_2="[ANY ADDRESS WITH FUNDS]"
+USER_PK_2="[ANY ADDRESS WITH FUNDS]"
 RPC=http://localhost:8547
 
 #TESTS
 CACHE_MANAGER_SIZE=2100000 # 2.1MB
 DUMMY_CONTRACTS_AMOUNT=6
+
+###########################################################
+
+# SEPOLIA EXAMPLE
+CACHE_MANAGER_ADDRESS=0x0c9043d042ab52cfa8d0207459260040cca54253
+ARB_WASM_CACHE_ADDRESS=0x0000000000000000000000000000000000000072
+CMA_ADDRESS="[DEPLOY AND DEFINE]"
+ARBPRE_ADD="[ANY ADDRESS WITH FUNDS]"
+ARBPRE_PK="[ANY ADDRESS WITH FUNDS]"
+USER_ADD="[ANY ADDRESS WITH FUNDS]"
+USER_PK="[ANY ADDRESS WITH FUNDS]"
+RPC=https://sepolia-rollup.arbitrum.io/rpc
