--- .env.sepolia
@@ -1,6 +1,0 @@
-ENVIRONMENT=testnet
-CACHE_MANAGER_ADDRESS=0x0c9043d042ab52cfa8d0207459260040cca54253
-ARB_WASM_CACHE_ADDRESS=0x0000000000000000000000000000000000000072
-ARBPRE_ADD=0x3f1Eae7D46d88F08fc2F8ed27FCb2AB183EB2d0E
-ARBPRE_PK=0xb6b15c8cb491557369f3c7d2c287b053eb229daa9c22138887752191c9520659
-RPC=https://sepolia-rollup.arbitrum.io/rpc
