+++ foundry.toml
@@ -1,0 +1,2 @@
+out = "build/out"
+cache_path = "build/forge-cache"
