+++ src/contracts/interfaces/ICacheManagerAutomation.sol
@@ -8,8 +8,19 @@ interface ICacheManagerAutomation {
     struct ContractConfig {
         address contractAddress;
         uint256 maxBid;
-        uint256 lastBid;
         bool enabled;
+    }
+    struct BidRequest {
+        address user;
+        address contractAddress;
+    }
+    struct BidResult {
+        bool shouldBid;
+        ContractConfig contractConfig;
+    }
+    struct UserContractsData {
+        address user;
+        ContractConfig[] contracts;
     }
 
     // Events
@@ -27,7 +38,6 @@ interface ICacheManagerAutomation {
         address indexed user,
         address indexed contractAddress,
         uint256 bidAmount,
-        uint256 minBid,
         uint256 maxBid,
         uint256 userBalance
     );
@@ -81,6 +91,18 @@ interface ICacheManagerAutomation {
         uint256 timestamp
     );
 
+    // Debug events
+    event DebugBidCheck(
+        address indexed user,
+        address indexed contractAddress,
+        string step
+    );
+    event DebugMinBidFetch(
+        address indexed contractAddress,
+        uint192 minBid,
+        bool success
+    );
+
     // Errors
     error InvalidAddress();
     error InvalidBid();
@@ -88,22 +110,34 @@ interface ICacheManagerAutomation {
     error ContractNotFound();
     error TooManyContracts();
     error ContractPaused();
+    error ContractAlreadyExists();
+    error ExceedsMaxUserFunds();
+    error InvalidFundAmount();
+    error TooManyBids();
 
     // Functions
-    function initialize(address _cacheManager, address _arbWasmCache) external;
-    function pause() external;
-    function unpause() external;
-    function getUserAddresses() external view returns (address[] memory);
-    function insertOrUpdateContract(
+    function insertContract(
         address _contract,
         uint256 _maxBid,
         bool _enabled
     ) external payable;
+    function updateContract(
+        address _contract,
+        uint256 _maxBid,
+        bool _enabled
+    ) external;
     function removeContract(address _contract) external;
     function removeAllContracts() external;
-    function setContractEnabled(address _contract, bool _enabled) external;
     function getUserContracts() external view returns (ContractConfig[] memory);
-    function getUserBalance() external view returns (uint256);
-    function withdrawBalance() external;
     function fundBalance() external payable;
+    function withdrawBalance() external;
+    function getUserBalance() external view returns (uint256);
+    function placeBids(BidRequest[] calldata _bidRequests) external;
+    function getContracts() external view returns (UserContractsData[] memory);
+    function getContractsPaginated(
+        uint256 offset,
+        uint256 limit
+    ) external view returns (UserContractsData[] memory userData, bool hasMore);
+    function getTotalUsersCount() external view returns (uint256);
+    function getUserAtIndex(uint256 index) external view returns (address);
 }
