+++ test/utils/deploy-dummy-wasm.sh
@@ -46,7 +46,7 @@ fi
 source "$ENV_FILE"
 
 # Move to contract folder
-cd test/mock-wasm-contracts/mock-contract
+cd test/utils/mock-wasm-contracts/mock-contract
 
 # Define the Rust source file
 CLEAN_RUST_FILE="src/clean_lib.rs"
@@ -84,7 +84,7 @@ cp $CLEAN_RUST_FILE $RUST_FILE
 cd ../../../
 
 # Create addresses.txt directory if it doesn't exist
-ADDRESSES_OUTPUT_FILE="test/scripts/addresses.txt"
+ADDRESSES_OUTPUT_FILE="test/utils/addresses.txt"
 mkdir -p "$(dirname "$ADDRESSES_OUTPUT_FILE")"
 
 # Append addresses to the addresses.txt file
