+++ test/utils/place-bids.sh
